requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
bitarray==2.9.2
blinker==1.8.2
Booktype==1.5
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==43.0.1
cytoolz==0.12.3
Django==5.1.1
et-xmlfile==1.1.0
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
google-api-core==1.34.1
google-auth==2.35.0
google-cloud-dialogflow==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.2
grpcio-status==1.48.2
hexbytes==1.2.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
msgspec==0.18.6
multidict==6.1.0
mysqlclient==2.2.4
numpy==2.0.2
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
parsimonious==0.10.0
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
PyMySQL==1.1.1
pyotp==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==15.1.0
redis==5.0.8
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rlp==4.0.1
rsa==4.9
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
SQLAlchemy==2.0.34
sqlparse==0.5.1
tldextract==5.1.2
toolz==0.12.1
twilio==9.3.1
types-requests==2.32.0.20240907
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
web3==7.2.0
websockets==13.0.1
Werkzeug==3.0.4
yarl==1.11.1
zope.event==5.0
zope.interface==7.0.3
